API tweaks. Added proper exceptions.

+++ src/main/java/com/mjabulani/privatePantry/api/ProductController.java
@@ -2,23 +2,30 @@ package com.mjabulani.privatePantry.api;
 
 import com.fasterxml.jackson.databind.cfg.MapperBuilder;
 import com.mjabulani.privatePantry.api.exception.ProductAlreadyExistsException;
+import com.mjabulani.privatePantry.api.exception.ProductDoesNotExist;
 import com.mjabulani.privatePantry.model.*;
 import com.mjabulani.privatePantry.repository.ProductRepository;
 import com.mjabulani.privatePantry.webclient.GptRequestBody;
 import com.mjabulani.privatePantry.webclient.GptResponse;
 import com.mjabulani.privatePantry.webclient.GptService;
 import com.mjabulani.privatePantry.webclient.Message;
+import jakarta.transaction.Transactional;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.yaml.snakeyaml.util.EnumUtils;
 import reactor.core.publisher.Mono;
 
 import java.util.*;
+
+import org.slf4j.Logger;
 
 
 @RestController
 class ProductController {
 
+    Logger logger = LoggerFactory.getLogger(ProductController.class);
     private final ProductRepository productRepository;
     private ProductCategory productCategory;
     public final GptService gptService;
@@ -30,28 +37,35 @@ class ProductController {
         this.gptService = gptService;
     }
 
-    // Get name of product by id
-    @GetMapping(
-            value = "products/name/{id}",
-            produces = "application/json")
-    String getProductName(@PathVariable int id) {
-        List<ProductEntity> p = productRepository.findById(id);
-        return p.get(0).getName();
-
-    }
+//    // Get name of product by id
+//    @GetMapping(
+//            value = "products/name/{id}",
+//            produces = "application/json")
+//    String getProductName(@PathVariable String id) {
+//        List<ProductEntity> p = productRepository.findById(id);
+//        return p.get(0).getName();
+//    }
 
     // Get product details by id
     @GetMapping(
             value = "products/{id}",
             produces = "application/json; charset=UTF-8")
-    Product getProductById(@PathVariable int id) {
-        ProductEntity p = productRepository.findById(id).get(0);
-        Amount amount = new Amount(p.getAmount(), p.getUnit());
-        return new Product(
-                p.getId(),
-                p.getName(),
-                p.getCategory(),
-                amount);
+    Product getProductById(@PathVariable String id) {
+        List<ProductEntity> productEntityList = productRepository.findById(id);
+        if (!productEntityList.isEmpty()) {
+            ProductEntity p = productEntityList.get(0);
+            return Product.builder()
+                    .id(p.getId())
+                    .name(p.getName())
+                    .category(p.getCategory())
+                    .amount(Amount.builder()
+                            .count(p.getAmount())
+                            .unit(p.getUnit())
+                            .build())
+                    .build();
+        } else {
+            throw new ProductDoesNotExist("Product with id = " + id + " does not exist!");
+        }
     }
 
 
@@ -64,15 +78,15 @@ class ProductController {
         List<Product> productList = new ArrayList<>();
         List<ProductEntity> products = productRepository.findAll();
         for (ProductEntity productEntity : products) {
-            Amount amount = new Amount(
-                    productEntity.getAmount(), productEntity.getUnit());
-
             productList.add(
                     Product.builder()
                             .id(productEntity.getId())
                             .name(productEntity.getName())
                             .category(productEntity.getCategory())
-                            .amount(amount)
+                            .amount(Amount.builder()
+                                    .count(productEntity.getAmount())
+                                    .unit(productEntity.getUnit())
+                                    .build())
                             .build());
         }
         return productList;
@@ -101,14 +115,27 @@ class ProductController {
             value = "products",
             produces = "application/json")
     @CrossOrigin(origins = "*")
-    ProductEntity addProduct(@RequestBody ProductAddRequest product) {
-        ProductEntity p = new ProductEntity(UUID.randomUUID().toString(), product.getName(), product.getCategory(), product.getAmount().getCount(), product.getAmount().getUnit());
+    Product addProduct(@RequestBody ProductAddRequest product) {
         if (productRepository.findByName(product.getName()) == null) {
-            productRepository.save(p);
+            ProductEntity pe = productRepository.save(ProductEntity.builder()
+                    .id(UUID.randomUUID().toString())
+                    .name(product.getName())
+                    .category(product.getCategory())
+                    .amount(product.getAmount().getCount())
+                    .unit(product.getAmount().getUnit())
+                    .build());
+            return Product.builder()
+                    .id(pe.getId())
+                    .name(pe.getName())
+                    .category(pe.getCategory())
+                    .amount(Amount.builder()
+                            .count(pe.getAmount())
+                            .unit(pe.getUnit())
+                            .build())
+                    .build();
         } else {
             throw new ProductAlreadyExistsException(product.getName() + " już istnieje");
         }
-        return p;
     }
 
     // Delete product by id
@@ -116,8 +143,13 @@ class ProductController {
             value = "products/{id}",
             produces = "application/json")
     @CrossOrigin(origins = "*")
-    void deleteProductById(@PathVariable int id) {
-        productRepository.deleteById(id);
+    @Transactional
+    void deleteProductById(@PathVariable String id) {
+        if (!productRepository.findById(id).isEmpty()) {
+            productRepository.deleteById(id);
+        } else {
+            throw new ProductDoesNotExist("Nie udało się usunąć produktu o id: " + id + ". Taki produkt nie istnieje!");
+        }
     }
 
 
@@ -126,19 +158,25 @@ class ProductController {
             value = "products/{id}",
             produces = "application/json")
     @CrossOrigin(origins = "*")
-    Product updateProduct(@PathVariable int id, @RequestBody ProductUpdate product) {
-        ProductEntity productToUpdate = productRepository.findById(id).get(0);
-        productToUpdate.setName(product.getName());
-        productToUpdate.setCategory(product.getCategory());
-        productToUpdate.setAmount(product.getAmount().getCount());
-        productToUpdate.setUnit(product.getAmount().getUnit());
-        productRepository.save(productToUpdate);
-        Amount amount = new Amount(productToUpdate.getAmount(), productToUpdate.getUnit());
-        return new Product(
-                productToUpdate.getId(),
-                productToUpdate.getName(),
-                productToUpdate.getCategory(),
-                amount);
+    Product updateProduct(@PathVariable String id, @RequestBody ProductUpdate product) {
+        List<ProductEntity> productEntityList = productRepository.findById(id);
+        if (!productEntityList.isEmpty()) {
+            ProductEntity productToUpdate = productEntityList.get(0);
+            productToUpdate.setName(product.getName());
+            productToUpdate.setCategory(product.getCategory());
+            productToUpdate.setAmount(product.getAmount().getCount());
+            productToUpdate.setUnit(product.getAmount().getUnit());
+            productRepository.save(productToUpdate);
+            Amount amount = new Amount(productToUpdate.getAmount(), productToUpdate.getUnit());
+            return new Product(
+                    productToUpdate.getId(),
+                    productToUpdate.getName(),
+                    productToUpdate.getCategory(),
+                    amount);
+        } else {
+            throw new ProductDoesNotExist("Product with id: " + id + " does not exist.");
+        }
+
     }
 
     @PostMapping(
